# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Insurance Reasoning Engine (IRE)",
    "description": "Insurance Reasoning Engine (IRE) - Clean Architecture FastAPI Backend Skeleton",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/agents/register": {
      "post": {
        "tags": [
          "Autonomous Agent Framework"
        ],
        "summary": "Register Agent",
        "operationId": "register_agent_api_v1_agents_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRegisterDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/run": {
      "post": {
        "tags": [
          "Autonomous Agent Framework"
        ],
        "summary": "Run Agent",
        "operationId": "run_agent_api_v1_agents_run_post",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/tasks/submit": {
      "post": {
        "tags": [
          "Autonomous Agent Framework"
        ],
        "summary": "Submit Task",
        "operationId": "submit_task_api_v1_agents_tasks_submit_post",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/tasks/{task_id}": {
      "get": {
        "tags": [
          "Autonomous Agent Framework"
        ],
        "summary": "Get Task Status",
        "operationId": "get_task_status_api_v1_agents_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/tools/register": {
      "post": {
        "tags": [
          "Autonomous Agent Framework"
        ],
        "summary": "Register Tool",
        "operationId": "register_tool_api_v1_agents_tools_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolRegisterDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/messages/send": {
      "post": {
        "tags": [
          "Autonomous Agent Framework"
        ],
        "summary": "Send Message",
        "operationId": "send_message_api_v1_agents_messages_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/messages/receive/{agent_id}": {
      "get": {
        "tags": [
          "Autonomous Agent Framework"
        ],
        "summary": "Receive Messages",
        "operationId": "receive_messages_api_v1_agents_messages_receive__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/shared-context": {
      "get": {
        "tags": [
          "Autonomous Agent Framework"
        ],
        "summary": "Get Shared Context",
        "operationId": "get_shared_context_api_v1_agents_shared_context_get",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentRegisterDTO": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role_description": {
            "type": "string",
            "title": "Role Description"
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt"
          },
          "primary_model": {
            "type": "string",
            "title": "Primary Model",
            "default": "gpt-4o"
          },
          "allowed_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Tools",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "name",
          "role_description",
          "system_prompt"
        ],
        "title": "AgentRegisterDTO"
      },
      "AgentRunDTO": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "default": "session_01"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds",
            "default": 60.0
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 200
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "goal"
        ],
        "title": "AgentRunDTO"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SendMessageDTO": {
        "properties": {
          "sender_agent_id": {
            "type": "string",
            "title": "Sender Agent Id"
          },
          "recipient_agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient Agent Id"
          },
          "message_type": {
            "type": "string",
            "title": "Message Type",
            "default": "REQUEST"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "content": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "sender_agent_id",
          "content"
        ],
        "title": "SendMessageDTO"
      },
...
      "ToolRegisterDTO": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "parameters_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters Schema",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "name",
          "description"
        ],
        "title": "ToolRegisterDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/hospitals/login"
          }
        }
      }
    }
  }
}